Add debugging facilities to RuleNode/TokenNode exposed via N-API (#908)
Closes #466 

## JSON serialization
Adds the `toJSON` method to the RuleNode/TokenNode classes with the
comment stating that it's unstable and for debugging purposes only. I
explicitly didn't type it (it's a string) as I didn't want to publicly
encode any assumptions about the shape of the data.

We can discuss further what we want to explore as part of having pure
data external interface (JSON) for Slang, whether that's exporting or
importing JSON, but I'd consider this a separate issue as it requires
more thought and design; the original issue, from what I understand, was
primarily about improving the debugging/inspection experience.

## Debugger
Adds hidden getter properties that is only eagerly evaluated by
debugger; since it's a getter function, it shouldn't add overhead to the
runtime but I have to admit, I didn't double-check that with numbers. It
adds both `__children` and `__text`, allowing to quickly explore the
tree structure recursively and also the enclosing text scope of the
node:

### vscode-js-debug (built-in to VS Code)

![image](https://github.com/NomicFoundation/slang/assets/3093213/193702fa-02b6-4853-9dcd-8af58469cf23)
### Chrome DevTools for Node

![image](https://github.com/NomicFoundation/slang/assets/3093213/a7b81836-ad7e-495c-aded-ebcc48f3bb6b)