Support Solidity 0.8.24
See <https://soliditylang.org/blog/2024/01/26/solidity-0.8.24-release-announcement>.

This update mostly introduced Cancun-related builtings, as well as a
global function.

This adds the new builtins but does not introduce the global function,
as global symbols will/should be added when we add the name resolution.
The name is also not reserved, so we're safe on that front.