fix: Ignore subsequent empty matches in a sequence when unwinding (#958)
Fixes #952 

When writing the original code, I expected to only encounter a `NoMatch`
being the token we unwound to, because I missed the case where the parse
sequence can have a child item that already attempts to recover but we
still expect more items in that sequence.

The code simply does not attempt to enforce the check and we simply
ignore subsequent empty matches until we hit the expected token in the
ancestor parser.