Allow only the first ellipsis matcher in an adjacent sequence to consume nodes
This is to avoid the weird query results that happen when there are two adjacent
ellipsis operators, where the results contain duplicates for every combination
of the nodes consumed by the ellipsis operators.

Adjacent here has a loose meaning, because if there are nil-matching operators
in between (eg. an optional operator), we need to propagate this information
forward.