unblock release pipeline
Release pipeline is [currently failing](https://github.com/NomicFoundation/slang/actions/runs/10281343571/job/28451407526) due to a tiny bug in `Command::run()`, where it returns a `Result<()>` even though it is panicking on error. This PR fixes that, and fixes all callsites to not expect a result.

Just a mechanical find/replace. The only functional change (the bug fix) is in `crates/infra/cli/src/toolchains/mkdocs/mod.rs` to use `Command::evaluate()` instead to check for the sub-process result.