Implement binding rules for control statements (#1076)
This PR adds support in the binding rules for:

- `if`/`else` statements
- `for` loops
- `while` and `do-while` loops
- `emit` statements and basic event definitions
- `try-catch` statements
- `revert` and basic error definitions
- `unchecked` blocks

It also includes a reorganization of the rules file to improve
readability and maintainability. Previously we were attempting to
maintain compatibility with strict evaluation mode which meant that
nodes and edges needed to be defined in previous rules. But since to
build the stack graph the lazy evaluation mode is used anyway, this
restriction can be dropped.

The file is now organized more or less hierarchically, with all the
rules affecting a node and its descendants grouped together.