Refactor: parents attribute allows defining topology
The `parents` attribute can now be applied both to definitions and
references (eg. to provide context for resolving `super` calls) and the
`selector` attribute can take the `alias` and `c3` values for now.

Scoring algorithm for C3 is not implemented yet, but should use the topology
defined via the `parents` attributes to linearise the definitions.