Add sample to inspect graph output and figure out injecting virtual methods
We want to reach all available virtual method definitions when setting a
compilation context for a contract, even when the calling site occurs in a base
contract. A call to `foo()` in a contract `Base` should be able to reach the
implementation in a sub-contract `Derived` even though there is not direct link
from `Base` to `Derived`.