Refactor bindings extra attributes and add import/export nodes
`import_nodes` and `export_node` will be used in post-processing to link the
virtual method definitions in a contract definition being compiled with all its
base contracts (recursively). This allows any call to a virtual method anywhere
in the hierarchy to be able to reach all the definitions available for that
method. This in combination with the ranking by linearisation should allow us to
choose the correct implementation when resolving a virtual call.