Implement Bindings context setting, which alters C3 linearisation
Setting the bindings context we indicate the bindings resolution that we're
working on a specific definition (eg. we're compiling a specific Contract type
in Solidity). This means that when resolving references, to account for virtual
method lookups, the universe of reachable definitions can now be expanded to
include implementations defined in more derived types in the hierarchy, up to
the set context.