Fix query engine `OneOrMoreMatcher` to skip over in-between trivia (#1133)
The `OneOrMoreMatcher` would stop matching whenever it encountered a
trivia item in a sequence. Since the parser may arbitrarily insert
trivia items anywhere, this meant that the matched results would be
incomplete.

This was found with this query from the Solidity bindings rules:

```
@specifier [InheritanceSpecifier [InheritanceTypes . @parents [_]+ .]]
```

Depending on the formatting of the parsed code, the parser would
generate the children of `InheritanceTypes` with interleaved trivia
items, eg. newlines if the inherited contracts/interfaces are enumerated
in separate lines. Because `[_]+` would stop on the first trivia, the
overall query would never match.