Fix binding structs as public getter results
A public getter will never return a complex data type, and a special getter
function is generated for arrays, mappings and structs. Arrays and mappings we
were already handling, but this commit makes that more explicit. For structs, we
bind the type of the first field. There is a caveat here: if the first field is
not a type that can be flattened and returned in a public getter we're still
binding it, instead of the first field that can be returned. But I don't see a
way to be 100% correct here.