Fix crash when a `YulPath` contains multiple `YulIdentifier`
This happens when there is special member access in some Yul identifier (like
`x.slot` or `x.offset`). I think this issue was introduced when unreserving the
`address` keyword since that changed the structure of `YulPath`. There is a
proper test case in #1149 already, but without this fix running Sanctuary with
existing rules crashes.