Add bindings test to Sanctuary (#1159)
This PR adds a step to Sanctuary testing to exercise bindings. We now
also consider a Sanctuary test to fail if any of the references found in
it cannot be resolved to one (or possibly more) definitions.

This also includes a small binding rules fix to avoid crashing if we
find a Solidity source file with an assembly block that has a `YulPath`
with more than one identifier (eg. a member access like `x.slot`). This
is properly tested in #1149, but without it running Sanctuary tests will
crash since there are contracts with the aforementioned contents.