Output graph files to `target` directory and improve Graphviz output (#1167)
The graph files `.mmd` and `.dot` files are not committed to the
repository and hence writing them to the source directory (inside the
`generated` folder) does not provide any benefit and even has some
drawbacks (eg. build errors when switching branches and test are no
longer present). This PR changes the output directory to be inside the
`target` directory in the `solidity_cargo_tests` crate directory.

These output files are opt-in via a `RENDER_GRAPHS` environment
variable. The performance impact of their generation is not too big, but
they are generally unnecessary unless we're debugging binding rules.

There are also some tweaks and added information to the `.dot` file
output to improve graph readability and make tracing nodes to rules
easier.