Fix `YulLabel` parsing (#1185)
Change the order that YulStatements get parsed so that
`YulVariableAssignmentStatement` gets tried before `YulLabel`. This
prevents and issue pre-0.5.0 where an assignment statement like this:

```solidity
assembly {
    foo:  = mload(0x80)
}
```

would get parsed as a label and then everything from `=` on would cause
an error. Now for versions before 0.5.5 this gets correctly parsed as a
variable assignment.

Closes NomicFoundation/slang#1182

---------

Co-authored-by: Omar Tawfik <15987992+OmarTawfik@users.noreply.github.com>