Precise error location in UNRECOGNIZED by attaching trivia to the expected nonterminal (#1172)
Fixes #841

In examples like the following:

```
// some trivia
unexpected
```

The parsing function was returning a terminal with `UNRECOGNIZED` kind
and the entire text inside. Since a terminal cannot have trivia
attached, we have to change the returning value to be a nonterminal. But
which one? For a moment I thought of a new specific `UNRECOGNIZED` kind,
but after discussing with Omar, it was clear that the better option was
to return the expected one: if we are parsing a `Statement`, then return
a `Statement` with the trivia and the `UNRECOGNIZED` terminal at the
end. This is realized in the `NoMatch` structure with a new field for
the optional expected nonterminal kind, which is essentially attached in
the `ParserResult::with_kind` method.

The commit history contains a previous attempt in which the trivia was
cached within the `NoMatch` struct. This added unnecessary complexity,
so I decided against it.

**Note1:** In a couple of places the code will use a
`ParserResult::no_match(<default values>)`. Noting there was a
`ParseResult::default()` function, I decided to use it instead, but I'm
not sure if this is the expected use of the default function.

**Note2:** I see that the error is one off at the end, probably
accounting EOF? But was already the case before, so there's no real
change there.