Summarize Sanctuary results (#1179)
This PR adds a result output when running in CI for `sanctuary` tests,
which are then passed as job outputs for each shard to the
`combinedResults` job that renders them in a table. We want to use this
to get a baseline and a benchmark for both correctness and performance
of the bindings rules.