Inject extension scopes while running the resolution algorithm (#1170)
This PR builds on top of #1149

This changes the way we handle extension scopes (ie. `using` directives)
in bindings. Instead of using the scope stack from the stack graph, we
hook into the resolution algorithm and inject new graph edges from nodes
designated as extension hooks (usually the source unit's lexical scope)
to extension scopes which are defined at each contract/library and
contain the definition nodes from `using` directives. This simplifies
the rules quite a bit and greatly improves performance, particularly in
the case of Solidity < 0.7.0 where `using` directives are inherited from
base contracts.

This PR also moves built-ins parsing and ingestion to `slang_solidity`
crate. Since the built-ins file needs to be pre-processed to transform
the symbols as to ensure no conflicts can occur with user code, adding
the built-ins requires a couple of manual steps that were replicated in
every construction of Solidity bindings API. By encapsulating this
functionality in the `slang_solidity` crate we remove a source of user
error and make it easier to make changes to the built-ins ingestion
code.