Change `ParseOutput` to return a `NonTerminal` instead of a `Node` (#1187)
Solves #1184
To go on top of #1172

This has a big impact in the failing examples, which now return the
expected `NonTerminal`.

This was acknowledgedly coded by playing the whack-a-mole with the type
system. I'm not so happy with the bunch of `Rc::clone()` that leaked
out; I will consider an alternative once we settle this is what we need
and there are no more pressing issues.

Currently doing a full CI run locally to find what more to fix. In the
meantime, I'll mark this as draft.