Modify the `infra test cargo` command to run more tests.
Previously, the command used to run tests was `cargo nextest run --workspace --all-features --lib --bins --examples --no-fail-fast`. This command ended up skipping the tests in the `metaslang` packages. We can run these as well by changing the command to `cargo nextest run --workspace --exclude solidity_testing_perf --all-features --all-targets --no-fail-fast`.

`--exclude solidity_testing_perf` is required because it tries to link with callgrind, which will not be available on all platforms. It's possible that this crash was the reason for the original restrictions in the first place, but I don't know for sure. We can think about different ways to conditionally run those tests in the future, if it's wanted.