Optimization: avoid extending complete partial paths a `@typeof` symbol is pushed next
This avoids a lot of extra work for the resolution when the identifier being
resolved is a variable that has a type. Since the definition is found but then
replace by the sequence `@typeof` and the type of the variable, which will never
resolve. This path *is* used when resolving for members of the type of the
variable.