Resolve references lazily (#1228)
This hopefully reduces resolution time when not all references (eg.
those in dependencies) need to be resolved, but requires the
`BindingGraph` to retain both the `StackGraph` and the `Resolver` (with
the database of partial path candidates and the partial paths arena).

This PR also includes an optimization to the resolution algorithm. When
selecting candidates to extend an already complete partial path, discard
those that would immediately push a `@typeof` symbol (eg. through the
path that replaces a variable by its type) since we know those will
never resolve. This can halve the number of evaluated partial paths (ie.
references that resolve to a variable/field definition) and provide a
very significant performance gain in some cases.