Change `Edge::label` from optional to required (#1205)
This involves a couple of changes outside of just changing the field
type:
1. `CursorIterator` must be able to produce a label for every node now.
This wasn't possible before because the root nodes of a cursor had no
parent, and so could not derive a label. I've changed `Cursor` in the
following ways to fix this:
* `Cursor::parent` is a new type, `Parent<T>`, instead of
`Option<Rc<PathAncestor<T>>>`. `Parent<T>` is an enum with three
variants - `None`, `Open(Rc<PathAncestor<T>>)`, and
`Closed(Rc<PathAncestor<T>>)`. A `Closed` parent is used when a cursor
is spawned from an arbitrary non-root node, and means that the parent
can be read but not traversed to. `None` is only used for the true root
node.
2. `KindTypes::EdgeLabel` is given a `Default` implementation, which for
now just returns the first `PredefinedLabel` variant. This is needed in
order to be able to set a "concrete" `EdgeLabel` value from the
`metaslang_cst` crate.
3. Added a `Root` variant to `PredefinedLabel`. This will be used when a
node is the true root node. For now this will be the default value for
`KindTypes::EdgeLabel`.

Closes #815

---------

Co-authored-by: Omar Tawfik <15987992+OmarTawfik@users.noreply.github.com>