move workflows outside devcontainer (#1247)
We cannot run the full CI in devcontainers, as we will run out of disk
space in GitHub Actions. Instead, we just validate the devcontainer is
built correctly by running a minimal smoke test in a single
`devcontainer` workflow, and move hte other workflows to run on the bare
GitHub Actions image.

Changes cherry-picked from #1225