Fix global `using` extensions propagation across multiple files
Closes #1249

Global extensions were being hooked directly to the source unit's lexical scope
instead of using a proper extensions scope. That meant it would only apply if
the type reference was located in the same file as the `using` directive and
would break eg. for functions in imported files returning the extended type.