Reduce size of CI builds (#1253)
The goal of this PR is to reduce the size of our build artifacts in CI
to avoid failures due to lack of disk space. This has been a sporadic
issue for a while, but it's become unavoidable starting with #1225,
which is constantly running out of space. That's because the new version
of JCO that it's bringing in is too large. We can build the project in
release mode in CI to compensate for this.

I've added some extra logic in `infra_cli` when running cargo commands
to append a `--release` flag when running in CI (indicated with the
environment variable `CI`). I also added a new flag to use in the
`infra` script (`scripts/bin/infra`), `SLANG_INFRA_RELEASE`, which will
cause the infra crates to be built in release mode too when set. With
both of these changes we can get the project size down to ~13 Gb with
either version of JCO. This keeps us under the [14 Gb
threshold](https://docs.github.com/en/actions/using-github-hosted-runners/using-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories).

---

Built with JCO rev. `d1b3344` (what `main` points to currently)

| Command | Total project size |
| --- | --- |
|  `infra ci` | 16 Gb |
| `CI=true infra ci` | 14 Gb |
| `CI=true SLANG_INFRA_RELEASE=true infra ci` | 13 Gb |

Built with JCO rev `b7818f3` (new rev that will come in once #1225 is
complete)

| Command | Total project size |
| --- | --- |
| `infra ci` | 17 Gb |
| `CI=true infra ci` | 14 Gb |
| `CI=true SLANG_INFRA_RELEASE=true infra ci` | 13 Gb |

---

I've also moved the generation of `public_api.txt` out of the unit tests
and in to a new step, `infra lint public-api`. I chose `infra lint`
because this is where the rustdocs are currently generated. This makes
the api docs generation a bit more visible and therefore easier to
debug, and also helps with unit test execution times.

Generating these docs causes debug artifacts to be generated - 1.2 Gb
worth, by my measurements. We might be able to reduce the overall
project size further by building this in release mode too, however the
crate we call which runs `cargo` to generate these does not provide an
API to do so. I believe that we could remove this crate entirely and
generate the rustdoc json file ourselves, but it didn't seem necessary
here and is outside the pure scope of this change so I left it as-is for
now.

---------

Co-authored-by: github-actions <github-actions@users.noreply.github.com>
Co-authored-by: Omar Tawfik <15987992+OmarTawfik@users.noreply.github.com>