Define built-ins contexts (for Yul built-in functions) and synthetically build the stack graph for the built-ins (#1203)
This PR introduces several changes to how built-ins are defined and
processed. We define different built-in contexts to separate built-in
symbols available in Solidity and Yul. Also, the part of the stack graph
that contains all definitions of built-ins is now built directly into
the stack graph, without generating a Solidity file and the processing
it as a system file. This allows much faster initialization of the
Bindings Graph and most symbol mangling is no longer needed since we're
not required to parse the built-ins file anymore.

In addition, all Yul built-ins are no longer defined in the grammar as
special keywords and instead are treated as normal identifiers and bound
to built-in definitions using the regular binding rules.

The API `add_system_file` was removed from `BindingGraphBuilder` and the
file descriptor renamed from `system` to `built-ins`. All code
previously required to transform built-in identifiers (ie. from `$` to
`%`) to avoid clashing with user code is no longer necessary and hence
removed.

Fix #1244: the changes for `address payable` are also included in this
PR.

---------

Co-authored-by: mjoerussell <mjoerussell@gmail.com>