Add larger example to the user guide (#1281)
This PR adds a much larger and moderately complex example to the user
guide. It's quite the jump from the previous examples, but it shows a
more advanced use case for Slang.

While writing it, at some point I tried using the AST API to implement
the decision logic on how to treat each kind of definition, in
particular with the contract. I quickly found out that that API is not
very usable for this kind of feature. The main flaw is that it's
difficult to retrieve a `Cursor` from an AST node or any of its members,
so this makes it cumbersome to mix with eg. the Binding Graph API.